The UART has been implemented and properly working. The initialization of peripherals has been moved from main to a separate file system_initialization.c. HAL_Driver/Inc/stm32f3xx_hal_def.h has been modified to let the system get many system definitions used by HAL.
