+++ Debug/objects.list
@@ -69,5 +69,6 @@
 "src/main.o"
 "src/stm32f3xx_it.o"
 "src/syscalls.o"
+"src/system_initialization.o"
 "src/system_stm32f3xx.o"
 "startup/startup_stm32f334x8.o"
