+++ NUCLEO-F334R8.xml
@@ -12,8 +12,8 @@
 <targetDefinitions>
   <board id="nucleo-f334r8">
     <name>NUCLEO-F334R8</name>
-    <dbgIF>JTAG</dbgIF>
     <dbgIF>SWD</dbgIF>
+    <dbgIF>JTAG</dbgIF>
     <dbgDEV>ST-Link</dbgDEV>
     <mcuId>stm32f334r8tx</mcuId>
   </board>
